Fixed plugin scripts being appended multiple times
