+++ src/index.js
@@ -3,8 +3,13 @@ import { render } from "react-dom";
 import { App } from "./components/App";
 import { extractID, isShareLink, REPLState } from "./state";
 
+import * as Babel from "@babel/standalone";
+import { plugins } from "./plugins-list";
+
 // css
 import "semantic-ui-less/semantic.less";
+
+window.babel = Babel;
 
 // If we want to be able to easily replace a codesandbox template via Define API later...
 // const BABEL_CONFIG = require("!raw-loader!../config.json");
@@ -80,7 +85,17 @@ async function getState() {
   return state === null ? defaultState : state;
 }
 
+function importDefaultPlugins() {
+  Object.keys(plugins).forEach(pluginName => {
+    const script = document.createElement("script");
+    script.src = plugins[pluginName].fileLocation;
+    script.async = false;
+    document.head.appendChild(script);
+  });
+}
+
 (async () => {
+  importDefaultPlugins();
   const state = await getState();
   render(
     <App
